flb: add third 'out_size' argument to co_create()
When a coroutine is created through co_create(), the caller pass a
required stack size, but due to alignments each architecture code
such as amd64, x86, ppx, etc might create a stack with some extra
bytes.

On development and debugging with memory checkers like Valgrind,
it will complain about stack switching so it requires the user
to register the stack addresses, for that we need to know what's the
real stack size created by the original libco, the new third parameter
'out_size' set's the real stack size so from the caller we can register
the stack properly.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>