aarch64: make sure co_switch_aarch64() stored in the text section
This is the second cut at aarch64.c to fix the segmentation fault
issue on ARMv8 platforms.

Instead of moving code around, I tweaked the assembly code to ask the
compiler to always store the function in ".text" section. With this,
we can be sure that co_switch_aarch64() won't be stored into a wrong
place and cause SIGSEGV.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>