core: use fiber backend instead of libco on Win64 build
This is a quick fix for the segementation fault bug I see occurring
on Win64 build.

    PS> .\bin\fluent-bit.exe
    ...
    [trace] [thread] destroy thread=00000297AB9BCD10 data=00000297AB9BCD30
    SIGSEGV

The root cause is not fully figured out yet, but it is very likely
that there is some Windows-specific memory management bug in arm64.c

I'll investigate the bug more later; Until any fix is found, we
need this in order to make fluent-bit work fine on Win64.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>